bot/requirements.txt
python-binance==1.0.19
websocket-client==1.8.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
tensorflow==2.16.2
streamlit==1.36.0
pyyaml==6.0.1
python-dotenv==1.0.1
requests>=2.31.0
ta>=0.10.0
backtrader>=1.9.78
vectorbt>=0.25.0
matplotlib>=3.8.0
plotly==5.22.0
python-telegram-bot>=20.7
openai>=1.6.0
transformers>=4.36.0
torch>=2.1.0
pytest>=7.4.0
pytest-mock>=3.12.0
schedule>=1.2.0
keras==3.4.1
cryptopanic==1.2.0
humanize==4.12.3 